Add style to contact section
