+++ src/nav.js
@@ -1,7 +1,5 @@
 import './../assets/zattix_store.png'
-import './../assets/culture.png'
-import './../assets/linters.png'
-import './../assets/todo.png'
+
 
 export default function displayNav(){
   const body = document.querySelector('body')
@@ -47,19 +45,10 @@ export default function displayNav(){
   })
 }
 
-
-
 function start() {
   document.querySelector('nav').innerHTML = ''
   displayNav()
-  const img1 = document.querySelector('.img1')
-  img1.src = './../assets/culture.png'
-  const img2 = document.querySelector('.img2')
-  img2.src = './../assets/zattix_store.png'
-  const img3 = document.querySelector('.img3')
-  img3.src = './../assets/todo.png'
-  const img4 = document.querySelector('.img4')
-  img4.src = './../assets/linters.png'
+
 }
